% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_tool.R
\name{noce_add_terms}
\alias{noce_add_terms}
\title{Add different term col}
\usage{
noce_add_terms(df)
}
\arguments{
\item{df}{noce data.frame}
}
\value{
df noce data.frame added varrious term cols
}
\description{
This function adds year, quarter of year, month term
to summaries value with various term
}
