% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noce_total.R
\name{noce_total_read_files}
\alias{noce_total_read_files}
\title{read total construction data}
\usage{
noce_total_read_files(file_path_v)
}
\arguments{
\item{file_path_v}{vector of excel files path}
}
\value{
tibble
}
\description{
This function reads excel files for number of
total construction.
}
